refactor(#130): clients 위치 수정
