--- apps/web/src/apis/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@horizon/api",
-  "version": "1.0.0",
-  "description": "Shared API client for Horizon Render",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "pnpm build --watch",
-    "typecheck": "tsc --noEmit",
-    "test": "echo 'no tests for api yet'",
-    "lint": "echo 'use root biome'",
-    "clean": "rm -rf dist"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "packageManager": "pnpm@10.12.4",
-  "type": "module",
-  "dependencies": {
-    "ky": "^1.10.0",
-    "zod": "^3.23.8"
-  },
-  "devDependencies": {
-    "@types/node": "^22.17.2",
-    "tsup": "^8.3.0",
-    "typescript": "^5.6.3"
-  }
-}
