--- apps/web/src/apis/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../../configs/tsconfig/library.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "lib": ["dom", "dom.iterable", "es2021"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
