+++ apps/web/src/clients/SignInContent/SignInContent.tsx
@@ -1,12 +1,12 @@
-import { useLogin } from '@/apis/auth/useAuth';
+import ContinueWithGoogle from '@/components/signin/ContinueWithGoogle/ContinueWithGoogle';
+import SecondaryAction from '@/components/signin/SecondaryAction/SecondaryAction';
+import { useLogin } from '@/hooks/auth/signin/useSignIn';
 import styled from '@emotion/styled';
 import { tokens } from '@horizon/tokens';
 import { Text, TextField } from '@horizon/ui';
 import { Flexbox } from '@horizon/utils';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
-import ContinueWithGoogle from '../../src/components/signin/ContinueWithGoogle/ContinueWithGoogle';
-import SecondaryAction from '../../src/components/signin/SecondaryAction/SecondaryAction';
 
 const SignInContent = () => {
   const [email, setEmail] = useState('');
