+++ apps/web/src/clients/SignUpContent/SignUpContent.tsx
@@ -1,12 +1,12 @@
+import EmailStep from '@/app/auth/steps/EmailStep';
+import PasswordStep from '@/app/auth/steps/PasswordStep';
+import UsernameStep from '@/app/auth/steps/UsernameStep';
+import VerificationStep from '@/app/auth/steps/VerificationStep';
+import { useSignUpHandlers } from '@/hooks/auth/signup/useSignUpHandlers';
 import styled from '@emotion/styled';
 import { tokens } from '@horizon/tokens';
 import { Icon, Text, TextField } from '@horizon/ui';
 import { Flexbox } from '@horizon/utils';
-import EmailStep from '../../src/app/auth/steps/EmailStep';
-import PasswordStep from '../../src/app/auth/steps/PasswordStep';
-import UsernameStep from '../../src/app/auth/steps/UsernameStep';
-import VerificationStep from '../../src/app/auth/steps/VerificationStep';
-import { useSignUpHandlers } from '../../src/hooks/useSignUpHandlers';
 
 const SignUpContent = () => {
   const {
@@ -84,7 +84,7 @@ const SignUpContent = () => {
                 onChange={(e) => {
                   setEmail(e.target.value);
                   if (errors.email) {
-                    setErrors((prev) => ({ ...prev, email: '' }));
+                    setErrors((prev: Record<string, string>) => ({ ...prev, email: '' }));
                   }
                 }}
                 error={!!errors.email}
@@ -109,7 +109,7 @@ const SignUpContent = () => {
             setVerificationCode={(value: string) => {
               setVerificationCode(value);
               if (errors.code) {
-                setErrors((prev) => ({ ...prev, code: '' }));
+                setErrors((prev: Record<string, string>) => ({ ...prev, code: '' }));
               }
             }}
             onSubmit={handleVerificationSubmit}
@@ -126,14 +126,14 @@ const SignUpContent = () => {
             setPassword={(value: string) => {
               setPassword(value);
               if (errors.password) {
-                setErrors((prev) => ({ ...prev, password: '' }));
+                setErrors((prev: Record<string, string>) => ({ ...prev, password: '' }));
               }
             }}
             confirmPassword={confirmPassword}
             setConfirmPassword={(value: string) => {
               setConfirmPassword(value);
               if (errors.confirmPassword) {
-                setErrors((prev) => ({ ...prev, confirmPassword: '' }));
+                setErrors((prev: Record<string, string>) => ({ ...prev, confirmPassword: '' }));
               }
             }}
             onSubmit={handlePasswordSubmit}
@@ -149,7 +149,7 @@ const SignUpContent = () => {
             setUsername={(value: string) => {
               setUsername(value);
               if (errors.username) {
-                setErrors((prev) => ({ ...prev, username: '' }));
+                setErrors((prev: Record<string, string>) => ({ ...prev, username: '' }));
               }
             }}
             onSubmit={handleUsernameSubmit}
